mxnet/requirements.txt
pytorch0.3.1
cv2
